+++ CHANGELOG.md
@@ -7,7 +7,7 @@ CHANGELOG
 
 * Updated `TransactionReport` to make the `ipAddress` parameter optional. Now
   the `tag` and at least one of the following paramters must be supplied:
-  `ipAddress`, `maxmindId`, `minfraudID`, `transactionID`.
+  `ipAddress`, `maxmindId`, `minfraudId`, `transactionId`.
 
 3.4.0 (2024-04-16)
 ------------------